CHE-84 Reorganized alumni controllers
