+++ server/controllers/alumniController/index.ts
@@ -1,0 +1,3 @@
+import getAllAlumniData from './getAllAlumniData/getAllAumniData';
+
+export { getAllAlumniData };
