+++ server/controllers/userController/index.ts
@@ -1,6 +1,6 @@
-import registerUser from './registerUser';
-import authUser from './authUser';
-import getUserById from './getUserById';
-import deleteUserByEmail from './deleteUserByEmail';
+import registerUser from './registerUser/registerUser';
+import authUser from './authUser/authUser';
+import getUserById from './getUserById/getUserById';
+import deleteUserByEmail from './deleteUserByEmail/deleteUserByEmail';
 
 export { registerUser, authUser, getUserById, deleteUserByEmail };
